+++ src/components/pages/TechnicalUserManagement/TechnicalUserTable.tsx
@@ -21,6 +21,7 @@
 import {
   IconButton,
   PageLoadingTable,
+  StatusTag,
 } from '@catena-x/portal-shared-components'
 import { useTranslation } from 'react-i18next'
 import ArrowForwardIcon from '@mui/icons-material/ArrowForward'
@@ -39,6 +40,14 @@ import Patterns from 'types/Patterns'
 interface FetchHookArgsType {
   statusFilter: string
   expr: string
+}
+
+enum ServiceAccountStatus {
+  ACTIVE = 'ACTIVE',
+  INACTIVE = 'INACTIVE',
+  PENDING = 'PENDING',
+  DELETED = 'DELETED',
+  PENDING_DELETION = 'PENDING_DELETION',
 }
 
 export const TechnicalUserTable = () => {
@@ -89,6 +98,20 @@ export const TechnicalUserTable = () => {
     },
   ]
 
+  const renderStatusColor = (status: string) => {
+    if (status.toLowerCase() === ServiceAccountStatus.ACTIVE.toLowerCase())
+      return 'confirmed'
+    else if (
+      status.toLowerCase() === ServiceAccountStatus.INACTIVE.toLowerCase()
+    )
+      return 'declined'
+    else if (
+      status.toLowerCase() === ServiceAccountStatus.DELETED.toLowerCase()
+    )
+      return 'deleted'
+    else return 'pending'
+  }
+
   return (
     <div style={{ paddingTop: '30px' }}>
       <PageLoadingTable<ServiceAccountListEntry, FetchHookArgsType>
@@ -126,7 +149,7 @@ export const TechnicalUserTable = () => {
           {
             field: 'serviceAccountType',
             headerName: t('global.field.type'),
-            flex: 1,
+            flex: 1.2,
           },
           {
             field: 'offer',
@@ -138,9 +161,22 @@ export const TechnicalUserTable = () => {
           {
             field: 'isOwner',
             headerName: t('global.field.owner'),
-            flex: 1,
+            flex: 0.8,
             valueGetter: ({ row }: { row: ServiceAccountListEntry }) =>
               row.isOwner ? 'Yes' : 'No',
+          },
+          {
+            field: 'status',
+            headerName: t('global.field.status'),
+            flex: 1.2,
+            renderCell: ({ row }: { row: ServiceAccountListEntry }) => (
+              <StatusTag
+                color={renderStatusColor(row.status)}
+                label={t(
+                  `content.usermanagement.technicalUser.status.${row.status}`
+                )}
+              />
+            ),
           },
           {
             field: 'details',